+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.2
 
 - General option `*`: dry run (stage or group does not affect the string, but the result can be printed), implies `:`.
+- Default regex for sort stages, `.+`.
 
 ## 0.8.1